Create anagrams.py
This is a really old anagram generator that I wrote. Ii don't want to look at it too much. It is not PEP-8 compliant. I even saw a global as I scanned through it. You need a dictionary. You can use /usr/share/dict/words, but it has lots of one and two letter words which clog up the results.

More info at:
http://mostlyhighperformance.blogspot.com/2012/01/generating-anagrams-efficient-and-easy.html